Gemfile: net-ssh =, not >= 3.2.0 because > versions may be different
As mentioned in the warning at the top of `Test.rb` we ended up
getting a little deeper into the internals of Net::SSH than I'm
totally comfortable with, so let's just lock the Gem version to 3.2.0
until we can try out other versions.

The particular worry here is that someone's got 4.x installed and that
would satisfy the >= 3.2.0 constraint.