SSH-AUTH.md Add a "References" section at the end
This version is just handy links at the end for all the
stuff referenced in-line. This makes it easier to get to
stuff related to this document without having to read
through it.

The links themselves are moved down, too, but this isn't
visible to someone reading the formatted document. We do not
interleave the links with the references because doing this
in an easily read way inserts extra paragraphs in the
Markdown-rendered text.